Error in symmetry excitation routine fixed.  Prior single excitations may be suspect.
Symmetry requirements for a single excitation used to be that the sym
prod of i and a (being occ and virt orbs) contain A1.  Now require that
(the sym prod of the remainder of orbs, having removed i)^2 x sym_i x
sym_a has A1.  This will not affect closed-shell determinants' singles.

Error in counting of singles discovered, with the number of excitations
in each symmetry class of singles being miscalcultated.  This may have
excluded some symmetry classes of excitations from being generated, and
would have led to errors in the excitations counts.  This is less likely
to be a problem with larger numbers of states, but small systems should
be checked.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@2722 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea