Added numerous options for UEG attenuated or cutoff exchange interactions.  Bug fixed in 2-vertex star of doubles which now gives excellent results.  Star diagonlisation now more efficient.   Will not yet work with complex orbitals.
EXCHANGE-CUTOFF [Rc]
   Sets cutoff distance Rc for the exchange electron-electron
potential.  If Rc  is not explicitly set, it will be set to be
equivalent to a sphere of the same volume.
dRc=(dOmega/(4*PI/3))**(1/3.d0)

EXCHANGE-DAMPING [Rc]
   Sets cutoff parameter Rc for attenuated potential
V(r)=erfc(r/Rc)/r).  If Rc  is not explicitly set, it will be set to be
equivalent to a sphere of the same volume.
dRc=(dOmega/(4*PI/3))**(1/3.d0)

    VERTEX STAR DIAG,POLY,POLYMAX  OLD,NEW
        POLY uses the special properties of the matrix to find the
roots of the  polynomial and uses them to calculate the relevant values.
This is order Ngrap h**2
        POLYMAX is POLY but only finds the highest root of the
polynomial, so i s order Ngraph.  It can be used when P is very large
(i.e. BETA is very large (s ay 40))
        POLYCONVERGE is poly but adds i out of N roots lambda_i such
that (N-i)  lamba_i**P < 10**-6, i.e. we evaluate enough roots such that
a very conservativ e error estimate of the contribution of the remaining
roots is negligible.
        OLD specifies to use a pre-generated list of determinants using
the exc itation routine version
               specified in EXCITATIONS OLD,NEW.
        NEW specifies to generate determinants on the fly without
storing them,  using the NEW excitation routine.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@4699 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea