StarMC now working in the case of changing one component at a time, either chosen stochastically, or sequentially. The stochastic implimentation doesn't seem too stable, though converges on the right answer as long as HAPP is sufficiently large - the number of times a component is selected to be updated, before the whole wavevector is updated.
A new file is also added, GraphMorph2.F90. This is not currently
working, or linked in to the rest of the program, but should provide a
way to do MC CI, which would be a new angle on the Graph Morphing. From
first impressions, it might well converge on the same result as normal
graph morphing, but the convergence rate will certainly be different.
This remains to be seen.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@9379 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea