Firstly, the Fixing of the nodal structure set by NODALCUTOFF is now enforced before the Annihilation of particles. This means that the number of particles at a site does not have to be zero if the signs are wrong, but rather the wrong signed particles will be destroyed before annihilation - they wil have no effect on the system.
Secondly, the MP1 wavefunction which is used for the fixed sign
approximation is now coded up exactly using hamiltonian matrix elements,
rather than the rho-element approximation.

Thirdly, if you use the nodal cutoff feature, you also get the
time-averaged energy of the wavefunction, calculated from the projection
of the doubles back onto the HF determinant. This seems to be more
stable than the shift as an estimator of the energy.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@9743 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea