Bug introduced in fixed-sign approximation in last commit, as fixed-sign approximation was being applied to the wrong vector.
New convention. If the nodal cutoff is negative, then the fixed-sign
approximation is applied BEFORE the annihilation of particles. This
means that there will still be a net positive weight on constrained
determinants, even if there were more negative particles.

Conversly, if the Nodal Cutoff is positive, then the fixed sign
approximation is applied AFTER annihilation of particles. This means
that if there is a net 'wrong' sign of particles, the weight on that
determinant is reset to zero.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@9745 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea