In FciMC.F90:
Added option: NOANNIHIL, which removes the sorting and annihilation step
from each cycle. This should be much faster, though remains to be fully
tested.
Also, all the data has been redone into pointers, which can be easily
swaped when needed.

In MCDets.F90:

Culling implimented, with same variables used as in FciMC.F90.
On exiting, the references to the excitation generators are set to zero
before deallocation to prevent an error message.
hHi0, and IC0, the Hamiltonian element to the HF determinant and the
excitation level of the particle, are added to ParticleData type. These
are not yet being used, but will be to calculate the energy later.


git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@9800 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea