Add vasp-neci communication routine to ifort GLOBALS.  vasp not currently compiled with any other compilers, so not added to the other GLOBALS files (yet).
Move the -r (relocatable) flag out of the explicit link line for the neci libaries and into the config files.  ifort (earlier than 10.0) is happier with the (more correct) -Wl,-r flag to specify the .a libraries are relocatable.

Create the targets for the cpmd (neci-cpmd.a) and vasp (neci-vasp.a) libraries respectively (and remove the neci.a target).

Create controlling makefile (see make help for options).



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@10217 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea