Fixed-node approximation implimented as originally published in D.F.B. ten Haaf, H.J.M. van Bemmel, J.M.J. van Leeuwen, W. van Saarloos, and D.M.Ceperley, Phys. Rev. B 51, 13039 (1995).
Results are rubbish, but then again, it was intended for use in
real-space DMC as a way to keep walkers from crossing the nodal
boundaries of their guiding wavefunction. It was unlikely to work here,
as we do not have localised regions of fixed sign. Death rate far too
high. Another way to simulate the effects of annihilation by increasing
on-site death needs to be thought of.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@10911 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea