Justification to the procedure name for the timing report.
It was the (intuitively named) adjustl function I wanted, rather than trim.  What's wrong with rstrip and lstrip?! I wish fortran was more like python.  :-)

(This is a duplication of the commit made in the rewrite branch: I just couldn't be bothered to do a merge for a change to a single word.)

git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@11049 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea