Changes to Serial FCIMC code:
Warning now printed if we are no annihilating
The output SumNoatHF has been changed to NoatHF, i.e. the instantaneous number at HF for that specific iteration which is printed out.
The same change has happened for NoatDoubs - no longer summed over the update cycle.

Changes to Parallel FCIMC code:

NoatHF, NoatDoubs and acceptance ratio now printed as for serial code.
The acceptance ratio is printed purely for the head node, though that should be good enough since they should all be very similar.
The option to start with a single particle at each node now available. STARTSINGLEPART as you would in serial.
Bugfix in the annihilation - VecSlot was not being initialised in one loop.
The sign of all the particles will now flip sign if the instantaneous population at HF becomes negative.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@11053 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea