Beginning of modularisation of Precalc. The data for precalcread and precalc is now in PrecalcData and is extensively commented. This should help with dependancies. Certain variables are no longer passed through to the main precalc routine and are used from modules.
I have removed the dependancy of NECI on StarMC and GraphMorph2 and put
these into the legacy code folder and explained why they are there. This
code will no longer be used.

New excitation weighting - STEPEXCITWEIGHTING. Here is the information
about it from the documentation:

**STEPEXCITWEIGHTING** [g_VMC_ExcitWeights(1) g_VMC_ExcitWeights(2)
G_VMC_EXCITWEIGHT]
This excitation weighting involves a step function between the HF
virtual and occupied electon manifold (i.e. step is at the chemical
potential)
When choosing an electron to move, the weighting to selecting it is increased
by 1 if the electron has energy above the chemical potential and by g_VMC_ExcitWeights(1,1)
if above. This occurs for both electrons. When choosing where to excite to, the
situation is reversed, and the probability of selecting it is increased by 1
if the electron is in the occupied manifold and g_VMC_ExcitWeights(2,1) if in
the occupied manifold. Bear in mind that the parameters are NOT probabilities.
If we are at a higher excitation level w.r.t. HF, then more electrons will be
in the virtual manifold, which will alter the normalisation, and mean that
when selecting electrons to excite, there will be an increasingly small
probability of selecting them from the occupied manifold. The opposite is true when
choosing where to put them.

U-weighting is the third parameter as before.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@11157 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea