A new FCIMC option. Here is the documentation on it:
**FIXHFDOUBSSHIFT** [DoubsShift]
    Default 0.D0

    An FCIMC Serial option. With this, the shift is fixed at a value given here, 
    but only for the HF and doubles. This will almost definitly give the wrong 
    answers for both the energy and the shift, but may be of use in equilibration 
    steps to maintain particle density at low excitations, before writing out the 
    data and letting the shift change.

Initial results are as expected for a DoubsShift value of 0 - HF and doubles populations increase
exponentially. HF particles cannot die (and are v unlikely to be
annihilated). Shift drops down, and presumably won't stop decreasing as
the population won't stop increasing. Energy estimator is surprisingly
good, though we must assume that as the shift drops lower and a larger
proportion of the total particles are on the HF + doubles, that the
'other' particles will pretty much die instantly, and we will be left
with convergence to the CISD result, rather than FCI.

It remains to be seen whether this is a productive equilibration
strategy.




git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@11351 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea