A slight change to the workings of WRITEDETE.
This will now write out a set of HISTEXCIT files each time the shift is
updated. Each file contains information about walkers from a different
excitation level, given by the filename. In these files is a histogram
of walkers in bins determined by the energy of the determinant, for all
walkers at that excitation level. This is calculated cumulativly over
the whole run.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@11436 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea