WRITEDETE now has two input parameters. First is the number of bins in the histogram (integer), and the second is the maximum determinant energy to histogram (a real).
Reading and writing of POPSFILEs in parallel is now functional,
although the energy estimators seem to be a little off - I'll look into
that. Scaling is still not functional in parallel.

Couple of other points in parallel - Errors are now fatal in MPI - they
were mistakenly set as returned. Various other slight changes in output
in parallel.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@11462 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea