Two bugfixes in the FCIMC parallel code - Both were only seen when REGENEXCITS was on, and caused segfaults in both the balancing routines and the culling routines. This was caused by CurrentExcits array still trying to be accessed, even though RegenExcits was on.
Also, with HFDOUBSSHIFT and STARTSINGLEPART, while we are still in the
single particle phase, the HF and doubles are still at the same shift as
the rest of the space. This will allow growth there are a higher rate.
Also, balancing is now possible from all particle numbers. This is to
minimise any impact that this has on growthrate values, which in turn
affect the change to the shift.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@11579 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea