Minor fix to allow compilation with ifort.
As ever, derived types are not consistently implemented across compilers. ;-)

However, using a pointer purely as an allocatable array in a derived type does not mean that it cannot be tested using associated.  When it is deallocated, it becomes null automatically (this does not apply to any other pointers pointing to it, however!).

git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@12664 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea