New excitation generators now interfaced with FCIMC in parallel. At the moment, excitation generators cannot be stored, though they are quick to generate. They can be used by specifying:
NONUNIFORMRANDEXCITS (CYCLETHRUORBS)

in the system block. The second optional keyword indicates that a chosen
orbital will be found by cycling through the orbitals of that symmetry,
rather than picking and discarding.

pDoubles, needed to give the probability of generating a double
excitation is now found from the proportion of HF excitations which are
double excitations in the FCIMC code.

Preliminary tests indicate a large speedup in generation of
random excitations by ~30%+ (and obviously no memory needed). They should be faster with larger basis set
sizes.

However, there have been a couple of problems.
Firstly, it is possible to generate determinants with no single
excitations allowed from them. If this happens, a double is now
generated and pDoub changed accordingly for that determinant.
Secondly, it seems that we can be put in positions where drawing and
redrawing forbidden orbitals can be pretty likely and so it dies. I will
try and code up something so that after a number of attempts, it changes
to cycling through the orbitals to choose one, which should be more
efficient for tricky cases.


SEPERATE NOTE (ahem AJWT):

It seems as though in the original random excitation generator, there
was a section of the code that ran over all symmetry pairs. This was
originally to write out the EXCITTYPES array, presumably for debugging,
but the write statements had long since been commented out. However, the
NSYM^2 operation was still there but redundant. This has now been
commented out too.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@13439 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea