MEMORYFAC and MEMORYFACEXCIT are now obsolete keywords. Two new keywords are introduced.
**MEMORYFACPART** [MemoryFacPart]
    Default 10.D0

    MemoryFacPart is the factor by which space will be made available for extra
    walkers compared to InitWalkers.

**MEMORYFACANNIHIL** [MemoryFacAnnihil]
    Default 10.D0

    MemoryFacAnnihil is a parallel FCIMC option - it is the factor by which space will be 
    made available for annihilation arrays compared to InitWalkers. This generally will need to be 
    larger than memoryfacpart, because the parallel annihilation may not be exactly load-balanced because of 
    non-uniformity in the wavevector and the hashing algorithm. This will tend to want to be larger 
    when it is running on more processors.

For parallel FCIMC calculations, it is important to have both options,
otherwise too much memory will be used. The way that these two are used
means that it is more streamlined and more memory has been saved from
this.

Other Changes:

Hubbard model calculations can now use the new excitation generators.
However, it can not cope with the symmetry and so it is important to
have NONUNIFORMRANDEXCITS NOSYMGEN specified. However, it is much
faster.

NOSYMMETRY is compatible with the new excitation generators. It was not
compatible with the old ones. However, this will not work with the
hubbard model.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@13483 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea