CONSTRUCTNATORBS option added in calc block.
After the number of iterations specified by 'EQUILSTEPS', this option 
calculates the one electron reduced density matrix (1-RDM) as a 
FCIMC calcation progresses.  
The 1-RDM is then diagonalised, giving linear combinations of the HF
orbitals that tend towards the natural orbitals of the system.  The
occupation numbers of these orbitals (e-values of the 1-RDM) are printed
in the OUTPUT file.





git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@13622 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea