Logging options added to RotateOrbs routine.
These are 

ROHISTOGRAM
which histograms the values of the four index integrals both before
rotation and after, and

ROFCIDUMP
which prints out an fcidump file of the final rotated integrals.

This rotate orbs routine is clearly still buggy as FCI and NECI runs in
the rotated orbital basis set do not converge to the correct energy yet.
Don't use this version.




git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@14605 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea