During the comparison of spawned particles with the main list of particles, we now transfer the particle accross if it is of the same sign. This is because all main arrays are sign-coherent with each other and so we know that it will never be annihilated. This should help with the scaling with number of processors too.
A note of warning though - this is the first time that we are actually
transferring particles between processors. Since they are compared with
the same processors determinants first, it shouldn't be a problem, as
the determinants on each processor should remain the same, but we might
run into load-balancing problems later on.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@14618 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea