RotoAnnihilation now only uses a single main array to hold walkers. Memory usage has been halved in one go.
NOTES: 

There seems to be some significant time taken in the annihilation
step which isn't in any of the timers. It seems to be associated with
attaching mpi buffers, and varies with the size of the buffer being
attached. I will try and isolate this and perhaps it'd be best to leave
the buffer attached rather than reattaching each iteration, but this has
its own problems.

The Insert/Remove routine is slower than what was being used before.
This is because we only have one array, and we have an extra N_s log
N_main operation. On the upside, this is only done once per iteration
and should hopefully not be too much of a problem in the future as I
test on larger systems.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@14765 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea