Large commit.
Bugfixes: Truncated FCIMC calculations were failing due to the excitation level only being calculated up to doubles. If truncated calculations are
          specified now, the exact excitation level is calculated.

Change to output: The average time per iteration is now printed out in the final column of the output file, rather than
                  the lowest occupied excitation level.

Large changes to the AnnihilateBetweenParts routine in preparation for the spawned lists becoming determinant
representations. Particles can now be specified multiple times on the same determinant. This will be useful for
when it comes to multiple spawns from the same particle. This shouldn't change the results at the moment.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@14931 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea