Hooray, we have converging energies from rotated orbitals.
New flag added to the System block.

**ROTATEDORBS**
Default false.

This indicates that the orbitals being read in are rotated (and not HF).
Therefore contributions from walkers on singly excited determinants must
be included in the energy calculations.

Also, option added to calc block.

**DEFINEDET [occupied orbitals]**
Default off.

This allows us to define a new reference determinant for the spawning.
The default is the same orbitals as before the rotation, and this seems
to be the most effective so far.

Finally, a binary file 'MOTransform' that contains the final transformation
matrix from the Rotation routine is now being printed out.  This will
hopefully be read into QChem to produce transformed cube files, however
it has not been tested yet.

Lots of commenting / documentation for all the Rotate stuff on its way...




git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@15067 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea