New option created.
**ERLOCALIZATION**
Default off.

Including this keyword in the system block sets the orbital localisation
to an Edmiston-Reudenberg localisation (i.e. maximising the <ii|ii>,
self interaction integrals).

With this comes a logging option.

**ERHISTOGRAM**
Default off.

This produces histograms of the <ii|ii> integrals before and after the
localisation.

Have also included requirements to only run over m and z of the same
symmetry.  





git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@15325 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea