Orbital rotation options added / changed slightly.
The original rotation in which |<ij|kl>|2 is minimised now requires the
system block keyword:

**OFFDIAGMIN**
This is to separate it from alternative rotations being developed such
as

**ERLOCALISATION**
performs an edmiston-reudenberg localisation as usual and 

**ERMAXOFFDIAGMIN** [ERWeight] [OffDiagWeight]

is some combination of the two above, each weighted
by a specified coefficient.  The weights have default values of 1.
However this final option is not yet working in this version.

Also, along with the **SEPARATEOCCVIRT** keyword,

**ROTATEOCCONLY**
default off.

and 

**ROTATEVIRTONLY**
default off.

have been added to rotate only the occupied or virtual and leave the
others as HF.  These can be used with symmetry on or symmetry off, but
not without the SEPARATEOCCVIRT keyword.




git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@15371 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea