Added parallel options to calc block:
**USEGUIDINGFUNCTION** [GuideDets]

and

**FINDGUIDINGFUNCTION** [InitGuideFuncParts]

The first writes out the GuideDets most populated determinants (and the
number of walkers with sign on these) to a file named GUIDINGFUNC.

The second reads in this file (scaling the number of walkers up to that
specified by InitGuideFuncParts), and uses it as a guiding function.
This guiding function may annihilate particles (and in the process be
annihilated itself), but may not die or spawn.
Its walkers are not included in the number of particles shown in the
FCIMCStats files, but they are included in the energy calculation.

These options are still not working quite right, and documentation is to
come.




git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@15988 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea