New parallel options added to the calc block:
**PRINTDOMINANTDETS** [NoDomDets] [MinExcLevel] [MaxExcitLevel]
At the end of a calculation, this finds the NoDomDets determinants with
the largest instantaneous population and excitation levels between
MinExcLevel and MaxExcitLevel.  
It then prints these to a file DOMINANTDETS

**SPAWNDOMINANTONLY**
This option reads in the DOMINANTDETS file and makes a list of these
dominant determinants.
These are to be the determinants that are allowed to be spawned on (for
the excitation levels chosen).
Spawning on any other determinant will be forbidden (but this part isn't
included in this commit).





git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@16009 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea