Bit added to PrintDominantDets that finds the normalisation deficiency that is expected by only including those determinants in the DOMINANTDETS file.
NormDef = 1 - [ sum_k->N_TR c_k^2 / sum_k->N_Tot c_k^2 ]
            
In other words the sum on the denominator goes over all determinants,
and the one on the numerator runs only over the dominant dets.
The default is to always find this parameter when writing out the 
DOMINANTDETS file.

Also changed the way the guiding function is read in, so that the 
parameter given with the key word **USEGUIDEFUNC** is now the 
number of walkers starting in the HF determinant of the guiding function
(per processor).
Rather than the total number in the guiding function over all
determinants.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@16271 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea