Large commit with bugfixes.
HPHF code now up and running. Plateaus seem to be reduced by ~2x, but
not all systems tested yet. Histogramming the wavefunction now works
with full diagonalisation, and now the ordering is done is such a way
that the determinants are ordered by excitation level first. This now
works for HPHF functions too.

The bug in the HPHF code is that the excitation generator can create the
same HPHF function again. The PGens need to be sorted for this, but I
have not currently forbidden the same excitations from being generated,
since this would mean non-trivial changes for the PGens. Instead, the
off-diagonal HPHF integral code will return zero if the same HPHF
function is sent in.

Also, if the HPHF function has an odd number of open-shell spatial
orbitals, the HPHF function is an anti-symmetric combination of
determinants, otherwise, it is symmetric. This has been put into the
HPHF integral generation scheme.

Many efficiency savings should now be put into the HPHF code.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@16610 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea