New parallel FCIMC logging option added.
**PRINTTRICONNECTIONS** [TriConMax] [NoTriConBins]
This prints out stats for sets of triangularly connected determinants.
Sign coherent triangles are ones where a walker on i spawns children
around the triangle and back onto i with the same sign.

The sum of the coherent and incoherent Hij x Hik x Hjk values are
printed in TriConnTotals, and the values are histogrammed at the end of
the calc in TriConnHistograms.

The first input parameter [TriConMax] is a real that determines the
range
of the histogram bins, and [NoTriConBins] is an integer to specify the
number of bins in that range.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@17383 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea