Another new parallel FCIMC logging option added. This goes along with the **PRINTTRICONNECTIONS** keyword.
**HISTTRICONNELEMENTS** [SingMax] [DoubMax] [NoBins]
This option histograms all the H elements involved in the triangular
connections of the above option. This will be a slightly different
distribution to those given by the **HISTENERGIES** option.  
The histogram range for the singles is specified to be -SingMax ->
+SingMax, or DoubMax for the doubles.  Both use the number of bins given
by NoBins.
Some stats are also written in the output file which give the average
size of the H elements involved.




git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@17541 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea