Firstly, EXPANDSPACE has been removed and replaced with EXPANDFULLSPACE. EXCITE should now always be used for trucated FCIMC calculations.
**EXPANDFULLSPACE** [iFullSpaceIter]
    Default=0

    This is a parallel FCIMC option. When this is set, the space initially is truncated at excitation level of ICIlevel,
    set by the value of the EXCITE parameter. If EXPANDFULLSPACE is set, then the
    system will continue to be truncated at the value set by EXCITE, but will
    expand to the full space after iteration iFullSpaceIter.
    Hopefully expanding the space in this way will allow quicker
    convergence, without needing to do this dynamically which may be difficult for
    long/queued jobs.

Also, all the EXPANDSPACE, SOFTEXIT and VARYSHIFT file searches have been removed and replaced by a CHANGEVARS file method.
If during the simulation, a CHANGEVARS file is created in one of the working directories, then the file is read
and many of the parameters of the run can be dynamically changed. The options in CHANGEVARS are as follows:

   EXCITE  XXX      Will change the excitation level of the simulation (< 0 or > NEl sets it to the full space.)
   SOFTEXIT         Will exit cleanly from the program
   WRITEPOPS        Will write a current popsfile
   VARYSHIFT        Will exit out of fixed shift phase
   TAU XXX          Will change the value of tau for the simulation
   DIAGSHIFT XXX    Will change the shift
   SHIFTDAMP XXX    Will change the damping parameter
   STEPSSHIFT XXX   Will change the length of the update cycle

Multiple options or single options can be specified at once. Since there is only one file now, it cuts out the time needed for searching for files and communication time between nodes with multiple options.

Also, it is easy to systematically add new parameters to vary dynamically without needing to write new routines or needing to search more files.

Hope this is useful!



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@17583 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea