New parallel FCIMC option added to INTEGRAL.
**FREEZEINNER** [No occupied inner to freeze],[No virtual inner to freeze]
This option freezes orbitals in the same way as **FREEZE** but, rather
than freezing the lowest energy occupied, and highest energy virtual,
it freezes the highest energy occupied and lowest energy virtual from
the fermi energy outwards.
This is very much still in the testing and clean up stage, but should
not affect the **FREEZE** option.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@17670 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea