**READCACHEINTS**     Default=.false.
    This means that the FCIDUMP file will be read in the integrals in it will be
    cached. This means that less space should be used for storage of the integrals,
    however, it will be slower since the integrals will need to be binary searched.

This should be working now, but only tested for small systems. Will test more now.

Note that I think I found a 'bug' introduced in the cacheing, whereby for real systems (but not RI), the <ii|jj> integrals
are calculated, rather than being looked up in UMat2D. Will check to see if this is general.




git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@17679 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea