New option added to the system block to go with the orbital rotation routine.
**READINTRANSMAT**
This reads in a file TRANSFORMMAT which contains a transformation
matrix, coefft1.  Without doing any sort of rotations, it uses this 
matrix to transform the 2 and 4 index integrals and print out a new 
dump file called ROFCIDUMP.
Using this keyword requires the **ROTATEORBS** keyword with two 
reals that in this case will not mean anything - this will 
eventually be changed so that the rotation options are in a separate
block.
Also, this only works with symmetry turned off at the moment.




git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@17695 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea