Few fairly large changes. I am now reasonably happy that CACHEFCIDUMP option is working as it should. More tests to follow.
1. The binary searching routine for the caching cannot find the last
element of an array. This is not too much of a problem if it can be
regenerated, but if the FCIDUMP file is cached, then it will return zero
- bit of a problem. This is now fixed. This should help as well with the
  normal use of the cache as less integrals will need to be regenerated,
  but this is slight.

2. All the caching of the FCIDUMP stuff has been fixed and moved into
UMatCache.F90. The cache is now explicitly read in in a contiguous
format, and then seperately later ordered and distributed by (j,l) pair
over the available slots. It will also precalculate the minimum memory
to do all this and output it. UMatEps will reduce this. Freezing still
needs to be fixed.

3. I have disabled caching new integrals after an integral miss when
tCacheFCIDUMPInts is set (which will occur if the integral is zero).

4. Direct access to UMAT2D from scr0 routine has been disabled. I may
come back to this when I work out exactly what is wrong with that and
reenable it, but it was giving weird results with the real integrals
from the FCIDUMP Cache.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@17710 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea