Allow the cache to work with spin-polarised systems. This is still done a little inefficiently, but hopefully I can improve that.
Rant: nBasisMax(2,3) is used for multiple purposes, which have major
conflicts now. It determines whether an integral should be
generated/looked up in cache, or read from UMAT, as well as whether it
is a spin-polarised/unpolarised system (sometimes called the ISS
variable). This scheme currently eliminates the
possibility for spin-polarised systems where you want to generate/look
up the integral in the cache. Now, whether the system is spin-polarised
or not is determined with the tspn flag, but this still isn't completely
satisfactory, and different places in the code deal with it is different
ways. I'll try and make this more consistent soon.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@17734 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea