New parallel FCIMC logging option to go with the orbital rotations.
**TRUNCROFCIDUMP** [NoFrozenVirt]
When using the MP2 variational density matrix to rotate the orbitals,
this option truncates the virtual space by removing the NoFrozenVirt
spin orbitals which have the smallest MP2VDM eigenvalues.
The ROFCIDUMP file printed only contains the unfrozen orbitals.
Still a bit of work to be done on the symmetry etc.




git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@17760 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea