New option:
**TRUNCATECAS** [OccCASOrbs] [VirtCASOrbs]

    This is a parallel FCIMC option, whereby the space will be truncated according to the specified CAS.
    The arguments indicate the active electrons, and then the number of active virtual orbitals.
    These values can be dynamically updated throughout the simulation via use of the CHANGEVARS facility.

Initial tests indicate that the convergence of the energy with increasing CAS space changes much slower than the normal
truncation by excitation level (as expected), however, the plateau variations as the CAS space is increased are interesting
and may be of use to get through the plateau faster.

Also added documentation on the **PRINTFCIMCPSI** logging option.

**PRINTFCIMCPSI**
    This works for parallel FCIMC. This will enumerate all excitations (up to the truncation level specified,
    or the full space if not specified), and then histogram the spawning run, writing out the final
    averaged wavefunction at the end.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@17931 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea