New option added to the orbital rotation in the system block.
**USECINATORBS**

This is similar to **USEMP2VDM** except the one electron reduced density
matrix is used to rotate the orbitals rather than the MP2VDM.
The rotation first requires a spawning calculation to find the
determinant coefficients.  This keyword turns on the required
histogramming, and performs the rotation at the end of the spawning
calculation.
The resulting orbitals are the approximate natural orbitals within the
specified truncation.  I.e. performing an excite 2 spawning calculation
will produce the CISD natural orbitals.
There is still a fair amount of testing and tidying to be done for this.




git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@17965 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea