Blocking error analysis added as default to parallel FciMC calculations. This can be turned off using:
**ERRORBLOCKING** OFF
In the logging block.

The default is for the blocking to begin when the HF population reaches
1000 over a shift update cycle.
This can be changed by using the following logging keyword:

**BLOCKINGSTARTHFPOP** [HFPopStartBlocking]
This will over ride the default and the blocking will start when the HF
pop over an update cycle reaches HFPopStartBlocking.

Alternatively using:
**BLOCKINGSTARTITER** [IterStartBlocking]
Will make the blocking start at iteration number IterStartBlocking,
rather than at a particular HF population.

Also, putting:
**STARTERRORBLOCKING**
in the CHANGEVARS file will cause the blocking to begin at that point.

A file BLOCKINGANALYSIS will the be printed at the end of the
calculation containing the error information for that run.
However this may be done in the middle of the calculation as well
by using **PRINTERRORBLOCKING** in the CHANGEVARS file.
This will print the error stats, but continue the calculation.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@18216 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea