Fix so NECI_H2 and NECIlong_Ne tests work.
Code was commented out in yersterday's commits such that the
single-particle energies were not being read in from FCIDUMP for
some calculations.  Code uncommented now.

This change needs to be checked, as there are other sections of code
which look like they should also be uncommented.

Please don't comment code out if it should be deleted!

git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@18326 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea