The symexcit3 file has now been added to objects, and the files using this array accordingly modified. This means that SymLabelList and SymLabelCounts now contain a new alpha/beta index so that these are always treated separately.
These new arrays SymLabelList2 and SymLabelCounts2 are then used in
symrandexcit2 - so **NONUNIFORMRANDEXITS** must be present in the system
block.
If these new arrays have been set up (which is always at the moment), a
flag tOddSymmetries is set to true, which results in a few changes to
FciMCPar - for instance in generating the double excitations for
**STARTMP1** calculations.

These changes shouldn't have affected anything except for the fact 
that if the symmetries are not read in in alpha/beta pairs, NECI can 
now deal with this.



git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@18404 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea